rename: group-filter.tsx 파일 폴더 변경
- src/components/molecules/group-filter 폴더로 이동
